// src/types/routes/submit/post-request.d.ts
/* eslint-disable */
/**
 * This file was automatically generated by json-schema-to-typescript.
 * DO NOT MODIFY IT BY HAND. Instead, modify the source JSONSchema file,
 * and run json-schema-to-typescript to regenerate this file.
 */

export interface PostRequest {
  course_name: string;
  assignment_name: string;
  tree: string;
  [k: string]: unknown;
}